backend IA/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
appscript==1.4.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
brotli==1.2.0
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dateparser==1.1.8
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
docx2pdf==0.1.8
docx2txt==0.9
docxtpl==0.20.2
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.1
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2026.1.0
g2pkk==0.1.2
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonlines==1.2.0
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.46.0
lxml==6.0.2
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.8.4
minio==7.2.20
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.15
mysql-connector-python==9.5.0
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba==0.63.1
numpy==1.26.4
openai==2.15.0
openai-whisper==20250625
packaging==25.0
pandas==1.5.3
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.5.1
playwright==1.57.0
pooch==1.8.2
preshed==3.0.12
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.0
PyJWT==2.11.0
pynndescent==0.6.0
pyparsing==3.3.1
pypdf==6.5.0
PyPDF2==3.0.1
pypdfium2==5.2.0
pyphen==0.17.2
pypinyin==0.55.0
pysbd==0.3.4
pytesseract==0.3.13
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
regex==2026.1.15
reportlab==4.4.7
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.11.4
sentence-transformers==5.2.3
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
SudachiDict-core==20251022
SudachiPy==0.6.10
sympy==1.13.1
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
trainer==0.0.36
transformers==5.2.0
TTS==0.22.0
typeguard==4.4.4
typer-slim==0.21.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.6.2
wasabi==1.1.3
weasel==0.4.3
weasyprint==67.0
webencodings==0.5.1
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
xlsxwriter==3.2.9
yarl==1.22.0
zopfli==0.4.0
supabase==2.28.0
